fstp: Add static to inline functions
From [1]
"This is needed to avoid a link error where the inline functions appear
missing at link time.
From c99 standard inline function should either be declared static or
have an extern instance in a c file for linking.
This fix is necessary to build with gcc 7; for some reason it was not
trigerred before."

[1] https://git.buildroot.net/buildroot/commit/?id=21133ada326c87627f7bdee4493d8086587c3cca

Signed-off-by: Romain Naour <romain.naour@gmail.com>
[Retrieved from:
https://git.buildroot.net/buildroot/tree/package/vde2/0002-fstp-Add-static-to-inline-functions.patch]
Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>